Bug 741986 - Partial solution to correcting too long sync interval when an Android device is only client.
This change makes SyncAdapter test for both an absolute "do not sync
before this time" (only set by server backoff requests) and a relative
"have we waited sufficiently long, based on the number of clients we
saw last sync".  This partially solves the problem on the bug in the
following case: if an Android device is the only client at some point,
we had set the absolute earliest next sync far in the future.  If we
then synced (for example, by force syncing) and a second client was
added, the absolute earliest next sync was not updated.  This fixes
that scenario.

To solve the problem in more cases, we should request a "clients only"
sync somewhat more frequently (so that we don't need to force sync),
say once an hour.

This should silently migrate (without fixing the issue).  The same
pref name is used for the earliest next sync time, but when client
syncs happen after whatever time was previously persisted before
migration, it will only be used when server backoff is requested.