Bug 773354 - Make TestAccountPickler only delete test Android Account objects; update test that passed for the wrong reason.
TestAccountPickler.testUnpickleAccountAlreadyExists passed because the
second Account created had the same parameters.  It was intended to
verify that if a Sync account already existed, we wouldn't unpickle
another one.  We don't (phew!), but that logic is in
SyncAccounts.syncAccountsExist.  I updated the test to verify the
correct flow doesn't create a second Account.