Bug 787502 - pre: Move credentials() from SyncStorageRequestDelegate to SyncStorageRequest.
Looking ahead to the auth changes coming in protocol 2.0, we can
re-use SyncStorageRequestDelegate if authentication-specific
configuration is moved out of the delegate and done at request
creation time.  This makes CredentialsSource the primary source for
credentials, which will make it easier to parameterize the auth scheme
in future.  Plus, the request delegate is intended for callbacks and
not really for authentication.