Bug 790931 - Broadcast when Sync Android Account is being deleted to many Firefox Apps.
The broadcast is secured with an Android signing-level permission and
specific to each Android Account type, so it is okay to include Sync
password and Sync key.

BroadcastReceiver.onReceive runs in the main thread and has a very
limited life-cycle, so we actually perform the work of cleaning up
after Account deletion in an associated IntentService.  (That work
being: delete the pickle file, wipe the client record from the remote
server, and purge the local shared prefs.)