Bug 814750 - Use ExecutorService in GlobalSession.advance.
By not nesting GlobalSession.advance calls, we hope to clear the stack
between stages and allow GC to release earlier stages.

The removed test verified only that GlobalSession calls
GlobalSessionCallback.handleSuccess on success.  This functionality is
better tested in the remaining TestGlobalSession file.

The removed test assumed that GlobalSession.start was synchronous --
which it *was*, but was an invalid assumption -- and fails with the
new, asynchronous, GlobalSession.advance/.start.  The test also did
not need to be on device.