Bug 815773 - Make GlobalSession.registerCommands() part of CommandProcessor singleton initialization.
As CommandProcessor states, there are two times when we need the
processor: to execute commands during a sync, and to store commands in
response to local user action.  CommandProcessor refuses to send
commands that have not been previously registered (probably the right
choice) and therefore we need to register all commands that we might
receive *and that we might send*.  Therefore, we move command
registration into the CommandProcessor initialization.

This tightly couples CommandProcessor and GlobalSession, which is
unfortunate, but could be revisited later.