Bug 833464 - Make forced syncs set the backoff interval.
Prior behaviour was that forced syncs *extended* the backoff interval.
This will mostly affect developers, since regular users do not force
syncs frequently, but one situation that could affect regular users is
when clients are added or removed from the Sync account and a sync is
manually forced.  In this case, if we have a long backoff interval
set, then a client is added, and then a sync is forced, we should set
a short backoff interval since we now have multiple clients.  The
prior behaviour would extend the backoff interval, keeping the
existing (incorrect) long interval.