No bug - Make android-sync APK start an activity when deployed to device.
Android does not deliver broadcast intents to a package before the
user has "started" it (see
http://commonsware.com/blog/2011/07/13/boot-completed-regression-confirmed.html
for discussion).  During development, it is awkward to "start" the
android-sync APK.  This commit adds a dummy
`BackgroundDeveloperActivity` and registers it, so that on deploy (at
least from Eclipse) Android "starts" the android-sync APK
automatically.  This makes testing `BroadcastReceiver` instances
smooth.