No bug - Add sharedpreferences-stub dependencies to Maven pom.xml files.
These dependencies are *not* required by maven's test runner, but they
are required to run tests under Eclipse.  This is due to differences in
how maven and Eclipse run Android tests (specifically, in how they
reference android.jar).  It's easiest to install the dependencies using
maven, so here they are.