Bug 967333 - Request sync when entering and leaving ConfirmAccountActivity.
This is the best we can do short of polling during the activity.  I
think if we are polling, we are honour-bound to show UI to indicate as
much to the user, and possibly let them control the poll.  Since we're
not going to land new UI for this now, the implemented solution will
have to suffice.