Bug 983856 - Message when syncing is disabled by Android.
This shows two new "error" (really, warning) states.  Both are shown
when the Firefox Account is correctly configured, and ready to sync, but
Android is preventing syncing automatically for some reason.  The first
is when Android is not syncing *any* Account
automatically (getMasterSyncAutomatically is false); the second is when
Android is not syncing *this Account
automatically (getSyncAutomatically) is false.