Bug 788688 - Part 3: Add "Device name" pref to Status activity.
In the edge case where what the user has entered (empty text) and what
is persisted (default client name) differ, Android does not update the
contents of the dialog's EditText correctly.  Removing and re-creating
all preferences is the only way I found to work around this; that's in
the next commit.