Bug 993134 - Part 4: Make fxa.TestAccountPickler robust.
This tested extracting data from an AndroidFxAccount before, and another
after, backed by different Android Accounts.  Unfortunately, one Account
replaced the other, but had the same name, leading to a race
condition (between reading and deleting the first account).  To address
this, we extract all the data before deleting, and then compare it to
the data from the newly pickled account.