Bug 956444 - Post: Persist fetch timestamp on 404.
I imported and modified your old code; this looks like an oversight from
there, but I want to call it out.  If we don't persist the timestamp,
then the stamp persisted to config is always -1, meaning we GET (and
404) the credentials on every sync.  You have code to compare
info/collections against the fetched timestamp so I think we should
persist even on 404.