Bug 964854 - Part 2: Maintain write-through memory cache of Firefox Account bundles.
This should avoid reads from the Android Account user data store, which
we theorize is buggy.  It trades those reads for the complexity of
maintaining and invalidating an in-memory cache, which has the potential
to avoid races and cache corruption.

There is no reliable way to determine if an Account has been
removed (and subsequently re-added), so we clear the cache entirely when
any Firefox Account is added.  We do this at the authenticator level,
which should be more inclusive than doing it at the AndroidFxAccount
level.

I put the cache itself in AndroidFxAccount, since that is where we have
been storing things associated to the Android Account object; but it
could just as well go in the authenticator.