Bug 1142596 - Part 3: Implement getAuthToken with token types of the form oauth::scope.
Be aware that there are two levels of token invalidation relevant here.
The first level is when a consumer uses an oauth token and gets a 401;
in this case, the consumer *must* call Android's own
invalidateAuthToken.  The second level is when the oauth client itself
gets a 401 trying to fetch an oauth token; in this case, the internal
state of the Firefox Account needs to be pushed back.