Bug BLOCKING - Part 1: Reduce scope of section that may set Account state.
The only place that might throw a TokenServerException is the token
server client code itself.  By handling such an exception earlier, we
reduce the scope of the section that may update the Firefox Account
state.  (This comes at the cost of threading AndroidFxAccount into
syncWithAssertion, but c'est la vie.)

This does not interact with the exist handling of 401s that we might see
from the storage endpoint.  Those 401s never generated
TokenServerExceptions; in fact, they were (essentially) ignored.  Since
we fetch a fresh token every Sync, what was (and is) expected is that
such 401s would be transient and fixed by authenticating with a fresher
token.

Test plan: manually verify that remotely changing the Firefox Account's
password while the device is in the Married state does the following:

1) uses the cached certificate to generate a local assertion;
2) the assertion produces a 401 from the TokenServerException, since the
certificate is no longer fresh;
3) the TokenServerException drives the Account state to Cohabiting;
4) the state machine discovers it cannot /sign, driving the Account
state to Separated.