Bug 1148504 - Part 2: Make updating Firefox Account state happen in a critical section.
It's worth noting that the two consumers of the shared state lock will
only race for a very short window -- essentially only when creating or
re-connecting an account.

That's because Reading List oauth tokens are long-lived and do not
expire (yet) in response to remote Account state changes, such as
updating the Account password.  So Sync and RL will race to initialize
the Account state; eventually RL will get an oauth token; and that token
will be cached forever until RL produces a 401 for the token or Android
expires the token.

Since Sync requests a token server token at the start of every sync, the
lock will be constantly exercised, but should never block.