Bug 1148029 - Disable Reading List sync when using custom endpoints.
Custom is interpreted to mean neither Mozilla production nor Mozilla
stage.  To ease the burden of testing stage, we sync the reading list
against stage reading list storage when the account is authing against
stage FxA.

Messaging this in some way would be nice but will have to wait for
follow-up.