Bug 1149226 - Initialize Reading List authority after upgrade.
The desired behaviour is: Reading List is syncing automatically if
Firefox Sync is syncing automatically.

Test plan:

0) Remove all Firefox Accounts.

1) Install old version.  Add a Firefox Account.  Ensure that Firefox
Sync is enabled.  Upgrade to new version.  Verify that Reading List is
enabled.

2) Install old version.  Add a Firefox Account.  Manually disable
Firefox Sync by unchecking the checkbox in Android Settings >
Accounts.  (This can not be done from the Firefox Account settings
activity; unchecking the 4 boxes there is not the same.)  Upgrade to new
version.  Verify that Reading List is disabled.

3) Install new version.  Add a new Firefox Account.  Verify that Reading
List is enabled.

(There are automated tests for these scenarios.)